requirements.txt
tqdm
streamlit
python-dotenv
google-genai